requirements.txt
appdirs==1.4.0
dj-database-url==0.3.0
dj-static==0.0.6
django==2.2.2
django-autocomplete-light==3.3.4
django-ajax-selects==1.5.0
django-appconf==1.0.2
django-constance==2.4.0
django-countries==5.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-googledrive-storage==1.3.4
django-picklefield==1.0.0
django-postgrespool==0.3.0
django-toolbelt==0.0.1
djangorestframework==3.9.2
enum34==1.1.6
google-api-python-client==1.6.4
gunicorn==19.5.0
httplib2==0.10.3
Markdown==2.6.9
oauth2client==4.1.3
packaging==16.8
Pillow==3.3.2
psycopg2==2.7.3.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pyparsing==2.1.10
python-dateutil==2.6.1
pytz==2015.2
reportlab==3.5.18
requests==2.20.0
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.3.0
sqlparse==0.2.4
static3==0.5.1
uritemplate==3.0.0
whitenoise==1.0.6
bootstrap-admin==0.4.3
